+++ firmware/controllers/math/engine_math.cpp
@@ -143,12 +143,14 @@ static const uint8_t order_1_3_7_2_6_5_4_8[] = { 1, 3, 7, 2, 6, 5, 4, 8 };
 static const uint8_t order_1_2_3_4_5_6_7_8[] = { 1, 2, 3, 4, 5, 6, 7, 8 };
 static const uint8_t order_1_5_4_8_6_3_7_2[] = { 1, 5, 4, 8, 6, 3, 7, 2 };
 static const uint8_t order_1_8_7_3_6_5_4_2[] = { 1, 8, 7, 3, 6, 5, 4, 2 };
+static const uint8_t order_1_5_4_8_3_7_2_6[] = { 1, 5, 4, 8, 3, 7, 2, 6 };
 
 // 9 cylinder
 static const uint8_t order_1_2_3_4_5_6_7_8_9[] = { 1, 2, 3, 4, 5, 6, 7, 8, 9 };
 
 // 10 cylinder
 static const uint8_t order_1_10_9_4_3_6_5_8_7_2[] = {1, 10, 9, 4, 3, 6, 5, 8, 7, 2};
+static const uint8_t order_1_6_5_10_2_7_3_8_4_9[] = {1, 6, 5, 10, 2, 7, 3, 8, 4, 9};
 
 // 12 cyliner
 static const uint8_t order_1_7_5_11_3_9_6_12_2_8_4_10[] = {1, 7, 5, 11, 3, 9, 6, 12, 2, 8, 4, 10};
@@ -201,6 +203,7 @@ static size_t getFiringOrderLength() {
 	case FO_1_2_3_4_5_6_7_8:
 	case FO_1_5_4_8_6_3_7_2:
 	case FO_1_8_7_3_6_5_4_2:
+	case FO_1_5_4_8_3_7_2_6:
 		return 8;
 
 // 9 cylinder radial
@@ -209,6 +212,7 @@ static size_t getFiringOrderLength() {
 
 // 10 cylinder
 	case FO_1_10_9_4_3_6_5_8_7_2:
+	case FO_1_6_5_10_2_7_3_8_4_9:
 		return 10;
 
 // 12 cylinder
@@ -221,9 +225,10 @@ static size_t getFiringOrderLength() {
 	case FO_1_14_9_4_7_12_15_6_13_8_3_16_11_2_5_10:
 		return 16;
 
-	default:
-		firmwareError(ObdCode::CUSTOM_OBD_UNKNOWN_FIRING_ORDER, "Invalid firing order: %d", engineConfiguration->firingOrder);
 	}
+
+	firmwareError(ObdCode::CUSTOM_OBD_UNKNOWN_FIRING_ORDER, "Invalid firing order: %d", engineConfiguration->firingOrder);
+
 	return 1;
 }
 
@@ -286,6 +291,8 @@ static const uint8_t* getFiringOrderTable()
 		return order_1_2_3_4_5_6_7_8;
 	case FO_1_5_4_8_6_3_7_2:
 		return order_1_5_4_8_6_3_7_2;
+	case FO_1_5_4_8_3_7_2_6:
+		return order_1_5_4_8_3_7_2_6;
 	case FO_1_8_7_3_6_5_4_2:
 		return order_1_8_7_3_6_5_4_2;
 
@@ -298,6 +305,8 @@ static const uint8_t* getFiringOrderTable()
 // 10 cylinder
 	case FO_1_10_9_4_3_6_5_8_7_2:
 		return order_1_10_9_4_3_6_5_8_7_2;
+	case FO_1_6_5_10_2_7_3_8_4_9:
+		return order_1_6_5_10_2_7_3_8_4_9;
 
 // 12 cylinder
 	case FO_1_7_5_11_3_9_6_12_2_8_4_10:
@@ -312,12 +321,11 @@ static const uint8_t* getFiringOrderTable()
 // do not ask
 	case FO_1_14_9_4_7_12_15_6_13_8_3_16_11_2_5_10:
 		return order_1_14_9_4_7_12_15_6_13_8_3_16_11_2_5_10;
-
-	default:
-		firmwareError(ObdCode::CUSTOM_OBD_UNKNOWN_FIRING_ORDER, "Invalid firing order: %d", engineConfiguration->firingOrder);
 	}
 
-	return NULL;
+	firmwareError(ObdCode::CUSTOM_OBD_UNKNOWN_FIRING_ORDER, "Invalid firing order: %d", engineConfiguration->firingOrder);
+
+	return nullptr;
 }
 
 /**