+++ firmware/CHANGELOG.md
@@ -38,6 +38,7 @@ or
  - Increased precision available for Lua tables
  - Add EGT values to CAN broadcast format #398
  - Add options to enable/disable optional CAN frames (cams, EGT so far)
+ - Add 1-5-4-8-3-7-2-6 and 1-6-5-10-2-7-3-8-4-9 firing orders
 
 ### Fixed
  - Improve performance with Lua CAN reception of a high volume of frames